requirements.txt
wechatpy>=1.8.0
requests>=2.25.0
fastapi>=0.68.0
uvicorn>=0.15.0
cryptography>=3.4.0
pydantic>=1.8.0
pycryptodome
